Implement a new settings system
It's still not a final solution, but a step in the right direction.
For now all QSettings-related logic is encapsulated in Core::Settings
class. It exposes just public variables, and that will be changed soon.
Also, lot's of settings keys have been changed to more logical values,
but some are still pending a better naming and a proper grouping.