Avoid creation of unnecessary objects
Qt containers return a deep copy when used in a loop with a call to a
member function such as QMap::keys() or QMap::values(const Key& key).
Using iterators or assigning the return value to a const object avoids
the copy, thanks to the implicit sharing feature of Qt.

Further, when arguments in a constructor are to be passed to a member's
constructor, taking the argument by value and then moving it can be
beneficial when the argument is an rvalue as it will avoid the call to a
copy constructor.