feat(ui): simplify combined docset download/extraction progress
This change replaces the progress bar with a plain label, that shows
numbers of docsets being downloaded and installed.

Also the Cancel button is now implemented as part of the dialog
button box. Because presently docset installation cannot be canceled,
the button stays enabled, but does nothing until all jobs are complete.

Fixes #1123.