feat: add option to hide menubar
Add a "View" submenu to the menubar, that allows the user hide
the menubar. Also add a keyboard shortcut (Ctrl + M) to toggle
the visibility of the menubar. This preference is saved, so it
can be restored once the user closes and reopens Zeal.

This preference is tipically only accessible through the menubar,
so don't add an entry for it in the "Preferences" window.